\name{ Fitting functions for GLD AFT regression }

\alias{survival.mle}
\alias{optim.survival.mle1}
\alias{optim.survival.mle2}
\alias{fun.simu.gld.lm.alt.surv}
\alias{GLD.lm.simu.surv}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{ This is a collection of functions designed to implement the fitting 
algorithms for GLD AFT regression in this package. }
\description{
These functions are for maintainers only and it is not designed for the users 
of this package.
}
\author{ Steve Su }
\value{ List of computation results for various fitting algorithms, not intended to be used by users of this package }
\note{ Please contact the author directly if you find a bug! }
\keyword{internal}