\name{ Fitting functions for GLD regression }

\alias{fun.fix.mean}
\alias{fun.simu.gld.lm.alt}
\alias{fun.simu.bias.correct.alt}
\alias{fun.model.lm.optim}
\alias{fun.model.lm.optim.Lmoment}
\alias{GLD.lm.simu}
\alias{fun.gld.all.vary.emp}
\alias{fun.gld.all.vary}
\alias{fun.gld.slope.fixed.int.vary.emp}
\alias{fun.gld.slope.fixed.int.vary}
\alias{fun.gld.slope.vary.int.fixed.emp}
\alias{fun.gld.slope.vary.int.fixed}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{ This is a collection of functions designed to implement the fitting 
algorithms for GLD regression in this package. }
\description{
These functions are for maintainers only and it is not designed for the users 
of this package.
}
\author{ Steve Su }
\note{ Please contact the author directly if you find a bug! }
\keyword{internal}